fix: create generatePartyHostRequest util for building PartyHostRequest payload
