+++ src/store/useHostStore.ts
@@ -1,7 +1,7 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
-interface PartyHostState {
+export interface PartyHostState {
   isGroup: boolean;
   groupName: string;
   name: string;
